+++ src/main/java/com/elmakers/mine/bukkit/utilities/RandomUtils.java
@@ -1,7 +1,5 @@
 package com.elmakers.mine.bukkit.utilities;
 
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Random;